chore:minor update to modal and button logic, WIP.
